printf() format warning fix
git-svn-id: https://svn.wxwidgets.org/svn/wx/wxWidgets/trunk@17181 c3d73ce0-8a6f-49c7-b76d-6d57e0e08775

+++ samples/menu/menu.cpp
@@ -419,7 +419,7 @@ wxMenu *MyFrame::CreateDummyMenu(wxString *title)
 
     if ( title )
     {
-        title->Printf(wxT("Dummy menu &%d"), ++m_countDummy);
+        title->Printf(wxT("Dummy menu &%u"), (unsigned)++m_countDummy);
     }
 
     return menu;
